+++ source_gen/lib/src/generator_for_annotation.dart
@@ -42,12 +42,25 @@ import 'type_checker.dart';
 /// may be helpful to check which elements have a given annotation.
 abstract class GeneratorForAnnotation<T> extends Generator {
   final bool throwOnUnresolved;
+  final String? inPackage;
+  final bool? inSdk;
 
   /// By default, this generator will throw if it encounters unresolved
   /// annotations. You can override this by setting [throwOnUnresolved] to
   /// `false`.
-  const GeneratorForAnnotation({this.throwOnUnresolved = true});
+  ///
+  /// With `source_gen` 4.0.0 this class will stop using mirrors for matching
+  /// annotations and will fall back to comparing the name of `T`. Pass
+  /// [inPackage] and [inSdk] to tighten the check; see [TypeChecker#typeNamed].
+  /// To use a custom annotation check, override [typeChecker].
+  const GeneratorForAnnotation({
+    this.throwOnUnresolved = true,
+    this.inPackage,
+    this.inSdk,
+  });
 
+  // This will switch to `typeNamed` in 4.0.0.
+  // ignore: deprecated_member_use_from_same_package
   TypeChecker get typeChecker => TypeChecker.fromRuntime(T);
 
   @override